HPACK decode on frame reception, not if/when/each-time a HEADERS frame is stringified.
Field compression is stateful.  Not to run HPACK decoding leads to header field
corruption.  To run HPACK decoding multiple times leads to header field corruption.

`h2get_hpack_decode` is now called from `h2get_mruby_conn_read` after reading
a frame of type HEADERS, PUSH_PROMISE or CONTINUATION, and the decompressed
list of header fields is stored in the list `field_lines` with the frame object.

The conversion to string now uses this list and it no longer touches the HPAC
state of the connection.